make uci parsing more resistant to errors

+++ src/misc.cpp
@@ -33,7 +33,7 @@ void PrintUciOutput(const int score, const int depth, const S_ThreadData* td, co
 	long  time = GetTimeMs() - td->info.starttime;
 	uint64_t nodes = td->info.nodes + getTotalNodes(options->Threads);
 
-	uint64_t nps = nodes / static_cast<uint64_t>((time + !time) * 1000);
+	uint64_t nps = nodes / (time + !time) * 1000;
 
 	if (score > -mate_value && score < -mate_score)
 		std::cout << "info score mate " << -(score + mate_value) / 2 << " depth " << depth << " seldepth " << td->info.seldepth << " multipv " << options->MultiPV << " nodes " << nodes <<
